requirements.txt
asgiref==3.8.1
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==41.0.7
Django==5.2.4
django-extensions==4.1
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
idna==3.10
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.0
pytz==2025.2
requests==2.31.0
sqlparse==0.5.3
typing_extensions==4.14.0
urllib3==2.5.0
uWSGI==2.0.26